python/CHAT2SQL-MODULE/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.7.1
pandas==2.2.2
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.1
